Split up main.c
