+++ wallaby_imu.h
@@ -1,0 +1,8 @@
+#include "stm32f4xx.h"
+
+void setupAccelMag();
+void setupGyro();
+
+void readAccel();
+void readMag();
+void readGyro();
