+++ wallaby_uart.h
@@ -1,0 +1,16 @@
+#ifndef WALLABY_UART_H_
+#define WALLABY_UART_H_
+
+
+// serial communication over external uart header
+void uart2_demo();
+
+
+// serial communication with main processor
+void uart3_demo();
+
+
+void setupUART2();
+void setupUART3();
+
+#endif
